Crate hopr_chain_api

Crate containing the API object for chain operations used by the HOPRd node.

Modules

config

errors

executors

Structs

ChannelEntry

Overall description of a channel

HoprChain

Represents all chain interactions exported to be used in the hopr-lib

SignificantChainEvent

Contains TX hash along with the Chain Event data. This could be used to pair up some events with Action. Each Action is typically concluded by a significant chain event.

Enums

HoprChainProcess

Functions

can_register_with_safe

Checks whether the node can be registered with the Safe in the NodeSafeRegistry

wait_for_funds

Waits until the given address is funded.

Type Aliases

DefaultHttpRequestor