hopr_chain_api

Struct HoprChain

pub struct HoprChain { /* private fields */ }

Represents all chain interactions exported to be used in the hopr-lib

NOTE: instead of creating a unified interface the HoprChain exports some functionality (e.g. the ChainActions as a referentially used) object. This behavior will be refactored and hidden behind a trait in the future implementations.

Implementations

impl HoprChain

pub fn new( me_onchain: ChainKeypair, chain_config: ChainNetworkConfig, node_db: HoprNodeDb, data_dir_path: &str, module_address: Address, contract_addresses: ContractAddresses, safe_address: Address, indexer_cfg: IndexerConfig, ) -> Result<Self>

pub async fn start(&self) -> Result<HashMap<HoprChainProcess, AbortHandle>>

Execute all processes of the HoprChain object.

This method will spawn the HoprChainProcess::Indexer and HoprChainProcess::OutgoingOnchainActionQueue processes and return join handles to the calling function.

pub fn me_onchain(&self) -> Address

pub async fn corrupted_channels(&self) -> Result<Vec<CorruptedChannelEntry>>

pub async fn get_indexer_state(&self) -> Result<IndexerStateInfo>